agent-eval/cli/pyproject.toml
[project]
name = "agent-eval"
version = "0.1.0"
description = "AgentEval CLI - Agent evaluation command-line tool"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Sean", email = "sean@example.com" }]
keywords = ["agent", "evaluation", "mlflow", "llm", "testing", "cli"]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
agent-eval = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
